requirements.txt
gunicorn==19.9.0
django==2.1
psycopg2==2.7.5
py-dateutil==2.2
xlrd=1.1.0
